Cleaning up project branch, removing unnecessary files
